refactor: co-locate date picker table composable
