--- src/components/VDatePicker/mixins/date-picker-table.ts
@@ -1,20 +1,0 @@
-import { defineComponent } from 'vue'
-
-// Composables
-import useDatePickerTable, { datePickerTableProps } from '../../../composables/useDatePickerTable'
-
-export default defineComponent({
-  name: 'date-picker-table',
-
-  props: {
-    ...datePickerTableProps,
-  },
-
-  setup (props, { emit }) {
-    const table = useDatePickerTable(props, emit)
-
-    return {
-      ...table,
-    }
-  },
-})
